Pass parameters rather than entire config
Bugfix: The `@config` was not used in `.verify_signed_time` and other
methods.

Also, the parameter passing is now uniformly not using `@config`. It
would not be pretty if we did try to always pass `@config` instead
because we should then create the `OpenSSL::PKCS7` and
`OpenSSL::X509::Certificate`s multiple times.